+++ lightgun/lib/OpenFIREPosition/OpenFIRE_Perspective_Advanced.h
@@ -88,7 +88,7 @@ private:
   inline float calculateQuadArea(float x0, float y0, float x1, float y1, float x2, float y2, float x3, float y3);
 
 public:
-  void warp(int x0, int y0, int x1, int y1, int x2, int y2, int x3, int y3, float dx0, float dy0, float dx1, float dy1, float dx2, float dy2, float dx3, float dy3);
+  void warp(float x0, float y0, float x1, float y1, float x2, float y2, float x3, float y3, float dx0, float dy0, float dx1, float dy1, float dx2, float dy2, float dx3, float dy3);
   void source(float adjustedX, float adjustedY);
   
   void setLensCorrection(float coefficientK1);
